Get it build ready
