+++ src/displayShows.js
@@ -1,7 +1,9 @@
 import makeCard from './makeCard.js';
-import showsLikes from './showsLikes.js';
+import mapLikes from './mapLikes.js';
+
 
 const displayShows = async () => {
+  const showsLikes = await mapLikes();
   const showsWrapper = document.querySelector('.shows-wrapper');
   const frag = document.createDocumentFragment();
   showsLikes.forEach((show) => {
@@ -9,5 +11,6 @@ const displayShows = async () => {
     frag.appendChild(card);
   });
   showsWrapper.appendChild(frag);
+  return showsLikes
 };
 export default displayShows;
