+++ src/index.html
@@ -16,7 +16,7 @@
       <h1>TV Shows Catalogue</h1>
     </div>
     <div id="select-genre">
-      <label for="genre">Select by Genre</label>
+      <label for="genre">Select by Genre:</label>
       <select id="genre" name="genre" disabled>
         <option value="drama">Drama</option>
         <option value="horror">Horror</option>
